pyproject.toml
[project]
name = "bigswarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "concrete-core[openai,requests]==0.0.7",
    "jinja2>=3.1.4",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"

[[tool.uv.index]]
url = "https://test.pypi.org/simple/"
default = true
